+++ install_tools.ps1
@@ -95,3 +95,34 @@ if (!(Test-Path "$INSTALL_DIR\webrtc\release\webrtc.lib")) {
   }
   Move-Item $BUILD_DIR\webrtc $INSTALL_DIR\webrtc
 }
+
+# CUDA
+
+if (!(Test-Path "$INSTALL_DIR\cuda\nvcc")) {
+  if ("$CUDA_VERSION" -eq "10.2") {
+    $_URL = "http://developer.download.nvidia.com/compute/cuda/10.2/Prod/local_installers/cuda_10.2.89_441.22_win10.exe"
+    $_FILE = "$BUILD_DIR\cuda_10.2.89_441.22_win10.exe"
+  } else {
+    # バージョンが増えたらこの分岐を増やしていく
+    throw "CUDA-$CUDA_VERSION URL not specified"
+  }
+
+  Push-Location $BUILD_DIR
+    if (!(Test-Path $_FILE)) {
+      Invoke-WebRequest -Uri $_URL -OutFile $_FILE
+    }
+  Pop-Location
+  if (Test-Path "$BUILD_DIR\cuda") {
+    Remove-Item "$BUILD_DIR\cuda" -Recurse -Force
+  }
+  mkdir $BUILD_DIR\cuda -Force
+  Push-Location $BUILD_DIR\cuda
+    # サイレントインストールとかせずに、単に展開だけして nvcc を利用する
+    7z x $_FILE
+    if (Test-Path "$INSTALL_DIR\cuda") {
+      Remove-Item $INSTALL_DIR\cuda -Recurse -Force
+    }
+    mkdir $INSTALL_DIR\cuda
+    Move-Item nvcc $INSTALL_DIR\cuda\nvcc
+  Pop-Location
+}